requirements.txt
dominate==2.1.16
Flask==0.10.1
Flask-Bootstrap==3.3.5.6
Flask-WTF==0.12
googlemaps==2.3
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pep8==1.6.2
python-google-places==1.0.1
requests==2.6.0
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
